+++ hybriddb/utils.py
@@ -8,7 +8,8 @@ from hybriddb.types import Column, SearchMode
 
 _SYSTEM_TABLES = {
     "_journal", "_schema", "_graph_nodes", "_graph_edges",
-    "_graph_sync", "_edge_rules",
+    "_graph_sync", "_edge_rules", "_versioned_tables",
+    "_version_checkpoints", "_chain_anchors",
 }
 
 JOURNAL_CAP = 50_000